+++ javabank-core/src/main/java/org/mbtest/javabank/fluent/IsBuilder.java
@@ -1,6 +1,6 @@
 package org.mbtest.javabank.fluent;
 
-import org.mbtest.javabank.http.core.Is;
+import org.mbtest.javabank.http.responses.Is;
 
 import java.io.File;
 import java.io.IOException;
