+++ javabank-core/src/test/java/org/mbtest/javabank/fluent/ImposterBuilderTest.java
@@ -2,11 +2,11 @@ package org.mbtest.javabank.fluent;
 
 import com.google.common.net.HttpHeaders;
 import org.junit.Test;
-import org.mbtest.javabank.http.core.Is;
 import org.mbtest.javabank.http.core.Stub;
 import org.mbtest.javabank.http.imposters.Imposter;
 import org.mbtest.javabank.http.predicates.Predicate;
 import org.mbtest.javabank.http.predicates.PredicateType;
+import org.mbtest.javabank.http.responses.Is;
 
 import static org.assertj.core.api.Assertions.assertThat;
