+++ javabank-core/src/test/java/org/mbtest/javabank/http/core/IsTest.java
@@ -6,7 +6,7 @@ import org.apache.http.HttpStatus;
 import org.json.simple.JSONObject;
 import org.junit.Before;
 import org.junit.Test;
-import org.mbtest.javabank.fluent.IsBuilder;
+import org.mbtest.javabank.http.responses.Is;
 
 import static org.assertj.core.api.Assertions.assertThat;
