+++ javabank-core/src/test/java/org/mbtest/javabank/http/core/StubTest.java
@@ -1,6 +1,7 @@
 package org.mbtest.javabank.http.core;
 
 import org.junit.Test;
+import org.mbtest.javabank.http.responses.Is;
 
 import static org.assertj.core.api.Assertions.assertThat;
